+++ state/api_stack.go
@@ -76,22 +76,3 @@ func (s *luaState) SetTop(idx int) {
 		}
 	}
 }
-
-func (s *luaState) PushNil() {
-	s.stack.push(nil)
-}
-func (s *luaState) PushBoolean(b bool) {
-	s.stack.push(b)
-}
-
-func (s *luaState) PushInteger(n int64) {
-	s.stack.push(n)
-}
-
-func (s *luaState) PushNumber(n float64) {
-	s.stack.push(n)
-}
-
-func (s *luaState) PushString(str string) {
-	s.stack.push(str)
-}
